+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:device_preview/device_preview.dart';
 import 'package:flutter/material.dart';
-import 'package:telebirr/pages/Loginpage.dart';
+import 'package:telebirr/screens/Loginpage.dart';
 import 'package:telebirr/screens/home_screen.dart';
 
 void main() {
